Move banner, license, author, uname functions from arch
- Create kos/banner.h header for banner, license, author functions
- Move banner/authors generation scripts from arch/dreamcast/kernel/ to kernel/
- Add arch/subarch description to the banner
- Add ARCH_NAME define, representing an arch's full name, to arch.h
- Move arch/dreamcast/kernel/uname.c to kernel/libc/koslib/
