+++ kernel/arch/dreamcast/include/arch/arch.h
@@ -85,6 +85,9 @@ unsigned HZ __depr("Please use the new THD_SCHED_HZ macro.") = THD_SCHED_HZ;
 /** \brief  Length of global symbol prefix in ELF files. */
 #define ELF_SYM_PREFIX_LEN  1
 
+/** \brief  Standard name for this arch. */
+#define ARCH_NAME           "Dreamcast"
+
 /** \brief  ELF class for this architecture. */
 #define ARCH_ELFCLASS       ELFCLASS32
 
@@ -300,47 +303,6 @@ void hardware_shutdown(void);
 */
 int hardware_sys_mode(int *region);
 
-/* These three aught to be in their own header file at some point, but for now,
-   they'll stay here. */
-
-/** \brief   Retrieve the banner printed at program initialization.
-    \ingroup attribution
-
-    This function retrieves the banner string that is printed at initialization
-    time by the kernel. This contains the version of KOS in use and basic
-    information about the environment in which it was compiled.
-
-    \return                 A pointer to the banner string.
-*/
-const char * __pure2 kos_get_banner(void);
-
-/** \brief   Retrieve the license information for the compiled copy of KOS.
-    \ingroup attribution
-
-    This function retrieves a string containing the license terms that the
-    version of KOS in use is distributed under. This can be used to easily add
-    information to your program to be displayed at runtime.
-
-    \return                 A pointer to the license terms.
-*/
-const char * __pure2 kos_get_license(void);
-
-/** \brief   Retrieve a list of authors and the dates of their contributions.
-    \ingroup attribution
-
-    This function retrieves the copyright information for the version of KOS in
-    use. This function can be used to add such information to the credits of
-    programs using KOS to give the appropriate credit to those that have worked
-    on KOS.
-
-    \remark
-    Remember, you do need to give credit where credit is due, and this is an
-    easy way to do so. ;-)
-
-    \return                 A pointer to the authors' copyright information.
-*/
-const char *__pure2 kos_get_authors(void);
-
 /** \brief   Dreamcast specific sleep mode function.
     \ingroup arch
 */
@@ -369,6 +331,15 @@ static inline bool arch_valid_address(uintptr_t ptr) {
 static inline bool arch_valid_text_address(uintptr_t ptr) {
     return ptr >= (uintptr_t)&_executable_start && ptr < (uintptr_t)&_etext;
 }
+
+/* The following functions are moved out of this header and are only provided for
+   compatibility reasons. Including any of them via this header is deprecated.
+*/
+
+/* Moved to <kos/banner.h> */
+const char * __pure2 kos_get_banner(void);
+const char * __pure2 kos_get_license(void);
+const char *__pure2 kos_get_authors(void);
 
 __END_DECLS
