+++ kernel/arch/dreamcast/kernel/init.c
@@ -10,6 +10,7 @@
 #include <string.h>
 #include <stdbool.h>
 #include <stdlib.h>
+#include <kos/banner.h>
 #include <kos/dbgio.h>
 #include <kos/dbglog.h>
 #include <kos/init.h>
