+++ kernel/banner.c
