+++ kernel/libc/koslib/Makefile
@@ -14,6 +14,6 @@ OBJS = abort.o byteorder.o memset2.o memset4.o memcpy2.o memcpy4.o \
 	creat.o sleep.o rmdir.o rename.o inet_pton.o inet_ntop.o \
 	inet_ntoa.o inet_aton.o poll.o select.o symlink.o readlink.o \
 	gethostbyname.o getaddrinfo.o dirfd.o nanosleep.o basename.o dirname.o \
-	sched_yield.o dup.o dup2.o pipe.o
+	sched_yield.o dup.o dup2.o pipe.o uname.o
 
 include $(KOS_BASE)/Makefile.prefab
